% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievement.R
\name{achievement}
\alias{achievement}
\title{Achievement API}
\source{
https://dev.battle.net/io-docs
}
\usage{
achievement(id)
}
\arguments{
\item{id}{id of the achievement}
}
\description{
Ahcievement API provides information on achievements.
}
\details{
\describe{
 \item{achievement}{This provides data about an individual achievement.}
 }
}

