% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\name{all_battlegroups}
\alias{all_battlegroups}
\alias{all_character_achievements}
\alias{all_character_classes}
\alias{all_character_races}
\alias{all_guild_achievements}
\alias{all_guild_perks}
\alias{all_guild_rewards}
\alias{all_item_classes}
\alias{all_pet_types}
\alias{all_talents}
\title{Data Resources}
\source{
https://dev.battle.net/io-docs
}
\usage{
all_battlegroups()

all_character_races()

all_character_classes()

all_character_achievements()

all_guild_rewards()

all_guild_perks()

all_guild_achievements()

all_item_classes()

all_talents()

all_pet_types()
}
\details{
\describe{
 \item{all_battlegroups}{The battlegroups data API provides the list of battlegroups for this region}
 \item{all_character_races}{The character races data API provides a list of each race and their associated faction, name, unique ID, and skin}
 \item{all_character_classes}{The character classes data API provides a list of character classes}
 \item{all_character_achievements}{The character achievements data API provides a list of all of the achievements that characters can earn as well as the category structure and hierarchy}
 \item{all_guild_rewards}{The guild rewards data API provides a list of all guild rewards}
 \item{all_guild_perks}{The guild perks data API provides a list of all guild perks}
 \item{all_guild_achievements}{The guild achievements data API provides a list of all of the achievements that guilds can earn as well as the category structure and hierarchy}
 \item{all_item_classes}{The item classes data API provides a list of item classes}
 \item{all_talents}{The talents data API provides a list of talents, specs and glyphs for each class}
 \item{all_pet_types}{The different bat pet types (including what they are strong and weak against)}
 }
}

