% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auction.R
\name{auction}
\alias{auction}
\title{Auction API}
\source{
https://dev.battle.net/io-docs
}
\usage{
auction(realm)
}
\arguments{
\item{realm}{}
}
\details{
\describe{
 \item{auction}{This API resource provides a per-realm list of recently generated auction house data dumps.}
 }
}

