% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss.R
\name{boss_list}
\alias{boss}
\alias{boss_list}
\title{Boss API}
\source{
https://dev.battle.net/io-docs
}
\usage{
boss_list()

boss(id)
}
\arguments{
\item{id}{id of specified boss}
}
\description{
Boss API provides information on bosses. A 'boss' in this context should be considered a boss encounter, which may include more than one NPC.
}
\details{
\describe{
 \item{boss_list}{A list of all supported bosses.}
 \item{boss}{The boss API provides information about bosses.}
 }
}

