% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_profile.R
\name{character_profile}
\alias{character_achievements}
\alias{character_appearance}
\alias{character_audit}
\alias{character_feed}
\alias{character_guild}
\alias{character_hunter_pets}
\alias{character_items}
\alias{character_mounts}
\alias{character_pet_slots}
\alias{character_pets}
\alias{character_professions}
\alias{character_profile}
\alias{character_progression}
\alias{character_pvp}
\alias{character_quests}
\alias{character_reputation}
\alias{character_statistics}
\alias{character_talents}
\alias{character_titles}
\title{Character Profile API}
\source{
https://dev.battle.net/io-docs
}
\usage{
character_profile(realm, characterName)

character_achievements(realm, characterName)

character_appearance(realm, characterName)

character_feed(realm, characterName)

character_guild(realm, characterName)

character_hunter_pets(realm, characterName)

character_items(realm, characterName)

character_mounts(realm, characterName)

character_pets(realm, characterName)

character_pet_slots(realm, characterName)

character_professions(realm, characterName)

character_progression(realm, characterName)

character_pvp(realm, characterName)

character_quests(realm, characterName)

character_reputation(realm, characterName)

character_statistics(realm, characterName)

character_talents(realm, characterName)

character_titles(realm, characterName)

character_audit(realm, characterName)
}
\arguments{
\item{realm}{name of the WoW Realm}

\item{characterName}{name of the character you want info on}
}
\description{
The Character Profile API is the primary way to access character information.
}
\details{
\describe{
 \item{character_achievements}{A map of achievement data including completion timestamps and criteria information}
 \item{character_appearance}{A map of a character's appearance settings such as which face texture they've selected and whether or not a healm is visible}
 \item{character_feed}{The activity feed of the character}
 \item{character_guild}{A summary of the guild that the character belongs to. If the character does not belong to a guild and this field is requested, this field will not be exposed}
 \item{character_hunter_pets}{A list of all of the combat pets obtained by the character}
 \item{character_items}{A list of items equipped by the character. Use of this field will also include the average item level and average item level equipped for the character}
 \item{character_mounts}{A list of all of the mounts obtained by the character}
 \item{character_pets}{A list of the battle pets obtained by the character}
 \item{character_pet_slots}{Data about the current battle pet slots on this characters account}
 \item{character_professions}{A list of the character's professions. Does not include class professions}
 \item{character_progression}{A list of raids and bosses indicating raid progression and completeness}
 \item{character_pvp}{A map of pvp information including arena team membership and rated battlegrounds information}
 \item{character_quests}{A list of quests completed by the character}
 \item{character_reputation}{A list of the factions that the character has an associated reputation with}
 \item{character_statistics}{A map of character statistics}
 \item{character_stats}{A map of character attributes and stats}
 \item{character_talents}{A list of talent structures}
 \item{character_titles}{A list of the titles obtained by the character including the currently selected title}
 \item{character_audit}{Raw character audit data that powers the character audit on the game site}

}
}

