% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auction.R
\name{download_auction}
\alias{download_auction}
\title{Download most recent auction data}
\usage{
download_auction(realm, destpath = tempdir())
}
\arguments{
\item{realm}{the name of the realm}

\item{destpath}{the path to save the file to}
}
\description{
helper function to download auction data
}

