% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\name{get_battlegroups}
\alias{get_battlegroups}
\alias{get_character_achievements}
\alias{get_character_classes}
\alias{get_character_races}
\alias{get_guild_achievements}
\alias{get_guild_perks}
\alias{get_guild_rewards}
\alias{get_item_classes}
\alias{get_pet_types}
\alias{get_talents}
\title{Data Resources}
\source{
https://dev.battle.net/io-docs
}
\usage{
get_battlegroups()

get_character_races()

get_character_classes()

get_character_achievements()

get_guild_rewards()

get_guild_perks()

get_guild_achievements()

get_item_classes()

get_talents()

get_pet_types()
}
\details{
\describe{
 \item{get_battlegroups}{The battlegroups data API provides the list of battlegroups for this region}
 \item{get_character_races}{The character races data API provides a list of each race and their associated faction, name, unique ID, and skin}
 \item{get_character_classes}{The character classes data API provides a list of character classes}
 \item{get_character_achievements}{The character achievements data API provides a list of all of the achievements that characters can earn as well as the category structure and hierarchy}
 \item{get_guild_rewards}{The guild rewards data API provides a list of all guild rewards}
 \item{get_guild_perks}{The guild perks data API provides a list of all guild perks}
 \item{get_guild_achievements}{The guild achievements data API provides a list of all of the achievements that guilds can earn as well as the category structure and hierarchy}
 \item{get_item_classes}{The item classes data API provides a list of item classes}
 \item{get_talents}{The talents data API provides a list of talents, specs and glyphs for each class}
 \item{get_pet_types}{The different bat pet types (including what they are strong and weak against)}
 }
}

