% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_profile.R
\name{get_character_profile}
\alias{get_achievements}
\alias{get_appearance}
\alias{get_audit}
\alias{get_character_profile}
\alias{get_character_talents}
\alias{get_feed}
\alias{get_guild}
\alias{get_hunter_pets}
\alias{get_items}
\alias{get_mounts}
\alias{get_pet_slots}
\alias{get_pets}
\alias{get_professions}
\alias{get_progression}
\alias{get_pvp}
\alias{get_quests}
\alias{get_reputation}
\alias{get_statistics}
\alias{get_titles}
\title{Character Profile}
\source{
https://dev.battle.net/io-docs
}
\usage{
get_character_profile(realm, characterName)

get_achievements(realm, characterName)

get_appearance(realm, characterName)

get_feed(realm, characterName)

get_guild(realm, characterName)

get_hunter_pets(realm, characterName)

get_items(realm, characterName)

get_mounts(realm, characterName)

get_pets(realm, characterName)

get_pet_slots(realm, characterName)

get_professions(realm, characterName)

get_progression(realm, characterName)

get_pvp(realm, characterName)

get_quests(realm, characterName)

get_reputation(realm, characterName)

get_statistics(realm, characterName)

get_character_talents(realm, characterName)

get_titles(realm, characterName)

get_audit(realm, characterName)
}
\arguments{
\item{realm}{name of the WoW Realm}

\item{characterName}{name of the character you want info on}
}
\details{
The Character Profile API is the primary way to access character information.
\describe{
 \item{get_achievements}{A map of achievement data including completion timestamps and criteria information}
 \item{get_appearance}{A map of a character's appearance settings such as which face texture they've selected and whether or not a healm is visible}
 \item{get_feed}{The activity feed of the character}
 \item{get_guild}{A summary of the guild that the character belongs to. If the character does not belong to a guild and this field is requested, this field will not be exposed}
 \item{get_hunter_pets}{A list of all of the combat pets obtained by the character}
 \item{get_items}{A list of items equipped by the character. Use of this field will also include the average item level and average item level equipped for the character}
 \item{get_mounts}{A list of all of the mounts obtained by the character}
 \item{get_pets}{A list of the battle pets obtained by the character}
 \item{get_pet_slots}{Data about the current battle pet slots on this characters account}
 \item{get_professions}{A list of the character's professions. Does not include class professions}
 \item{get_progression}{A list of raids and bosses indicating raid progression and completeness}
 \item{get_pvp}{A map of pvp information including arena team membership and rated battlegrounds information}
 \item{get_quests}{A list of quests completed by the character}
 \item{get_reputation}{A list of the factions that the character has an associated reputation with}
 \item{get_statistics}{A map of character statistics}
 \item{get_stats}{A map of character attributes and stats}
 \item{get_character_talents}{A list of talent structures}
 \item{get_titles}{A list of the titles obtained by the character including the currently selected title}
 \item{get_audit}{Raw character audit data that powers the character audit on the game site}

}
}

