% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{item}
\alias{item}
\alias{item_set}
\title{Item API}
\source{
https://dev.battle.net/io-docs
}
\usage{
item(itemid)

item_set(setid)
}
\arguments{
\item{itemid}{id of the item}

\item{setid}{id of the set}
}
\description{
Item API provided detailed item information.
}
\details{
\describe{
 \item{item}{The item API provides detailed item information. This includes item set information if this item is part of a set}
 \item{item_set}{The item API provides detailed item information. This includes item set information if this item is part of a set}
 }
}

