% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quest.R
\name{quest}
\alias{quest}
\title{Quest API}
\source{
https://dev.battle.net/io-docs
}
\usage{
quest(id)
}
\arguments{
\item{id}{id of specified quest}
}
\details{
\describe{
 \item{quest}{Retrieve metadata for a given quest}
 }
}

